AgenticOS/requirements.txt
google-genai
web3
flask
flask-cors
python-dotenv